Remove base repository class to just make things simpler
